requirements.txt
apimatic-core==0.1.4
apimatic-core-interfaces==0.1.1
apimatic-requests-client-adapter==0.1.1
asgiref==3.6.0
CacheControl==0.12.11
certifi==2022.12.7
charset-normalizer==3.0.1
deprecation==2.1.0
Django==4.1.7
django-address==0.2.8
django-environ==0.9.0
django-phonenumber-field==7.0.2
enum34==1.1.10
human-readable==1.3.2
idna==3.4
jsonpickle==1.5.2
msgpack==1.0.4
packaging==23.0
phonenumbers==8.13.5
Pillow==9.4.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
squareup==25.1.0.20230119
stripe==5.1.1
urllib3==1.26.14
